Refactor validation actions to support int type and enhance error messaging
- Updated numberAction to handle int instead of int64 for validation.
- Refactored validation functions (Min, Max, Gt, Gte, IsPositive, IsNegative, IsZero, Lt, Lte, CustomNumber) to accept int parameters.
- Modified IntPipeManager and IntPipeAction interfaces to work with int values.
- Enhanced error message handling in stringAction to provide context-specific messages.
- Introduced new regex validation functions for UUIDs, Bitcoin addresses, and other formats.
- Added new string validation functions (Pattern, MaxLength, MinLength, IsAlpha, IsDecimal, IsEmail, etc.) with improved error messaging.
- Updated Schema interface to separate Validate and ValidateAll methods for better error handling.
- Created a new regex.go file for pre-compiled regex patterns to improve performance.

+++ lib/v/int_pipe.go
@@ -1,26 +1,26 @@
 package v
 
-// IntPipeManager manages the validation pipeline for int64 values.
+// IntPipeManager manages the validation pipeline for int values.
 type IntPipeManager struct {
 	actions []IntPipeAction
-	value   int64
+	value   int
 	key     string
 	error   error
 }
 
-// IntPipeAction defines the interface for int64 validation actions.
-// Each action can run validation logic on an int64 value and return an error if validation fails.
+// IntPipeAction defines the interface for int validation actions.
+// Each action can run validation logic on an int value and return an error if validation fails.
 type IntPipeAction interface {
-	Run(v int64) error
+	Run(v int) error
 }
 
-// NewNumberPipe creates a new validation pipe for int64 values.
+// NewNumberPipe creates a new validation pipe for int values.
 // The pipe executes the provided actions in sequence during validation.
 //
 // Example:
 //
 //	pipe := NewNumberPipe(42, Min(0), Max(100))
-func NewNumberPipe(value int64, actions ...IntPipeAction) Pipe {
+func NewNumberPipe(value int, actions ...IntPipeAction) Pipe {
 	return &IntPipeManager{
 		value:   value,
 		actions: actions,
